Added Discovery_Used table
This table represents the aggregation between the discovery relationship, and the instrument used in the discovery.

The discovery table has the primary key of an astronomer id, instrument id, and celestial body id. It also has a date of discovery attribute.

The model for this table describes that it belongs to an instrument, and the instrument model was updated to say that it has many discoveries. This allows for easy querying of all discoveries corresponding to an instrument. Similarly, this makes it easy to query the instrument corresponding to a discovery.